create profile and icon

+++ front/src/pages/showUser.tsx
@@ -1,7 +1,8 @@
 import axios from "axios";
+import { useState } from "react";
 import { Link } from "react-router-dom";
 
-function showTweet(props: {
+function showUser(props: {
   userID: number;
   username: string;
   displayName: string;
@@ -9,38 +10,39 @@ function showTweet(props: {
   isFollowed: boolean;
   isFollowing: boolean;
 }) {
+  const [isFollowing, setIsFollowing] = useState(props.isFollowing);
   return (
     <div className="user">
-      <div>名前：{props.displayName}</div>
       <div>
-        ユーザー名：
+        <Link to={`/users/${props.username}`} key={props.username}>
+          {props.displayName}
+        </Link>
+      </div>
+      <div>
         <Link to={`/users/${props.username}`} key={props.username}>
           @{props.username}
         </Link>
       </div>
-      <div>フォローされてるか：{props.isFollowed ? 1 : 0}</div>
-      <div>フォローしているか：{props.isFollowing ? 1 : 0}</div>
-      {props.bio.Valid && <div>自己紹介：{props.bio.String}</div>}
+      {props.isFollowed && <div>フォローされています</div>}
       <div>
         <button
           type="submit"
           onClick={() => {
-            axios.post(`/api/users/${props.username}/follows`);
+            if (isFollowing) {
+              axios.delete(`/api/users/${props.username}/follows`);
+              setIsFollowing(!isFollowing);
+            } else {
+              axios.post(`/api/users/${props.username}/follows`);
+              setIsFollowing(!isFollowing);
+            }
           }}
         >
-          フォローする
-        </button>
-        <button
-          type="submit"
-          onClick={() => {
-            axios.delete(`/api/users/${props.username}/follows`);
-          }}
-        >
-          フォロー解除する
+          {isFollowing ? <div>フォロー中</div> : <div>フォローする</div>}
         </button>
       </div>
+      {props.bio.Valid && <div>{props.bio.String}</div>}
     </div>
   );
 }
 
-export default showTweet;
+export default showUser;
